Started Daily Report Script. Data Prep Overhual
This is really one of those, "I want to do some major refactoring
and better do that in another branch. oh. I haven't done a commit in
a while. whoops" commits.

At this time I've reorganized the WCHD data such that all the stats are
computed by python and only raw data is stored in the google sheet.
Sheets were renamed so update_data.sh needed to be corrected .
This also created the need for a daily check in script to look at
day-to-day stats that were previously computed in sheets. That's the
script covid-warren-daily.py. The perparation of the WCHD data takes
place in cvdataanalysis.py.  Visualization routines needed to be
updated to account for new dataframe structures.

Some work has been done on preparing and reorganizing the usafacts data
set to facilitate wider analysis and make integrating that data with
WCHD data a bit easier. This is a work in progress, shouldn't be on
master, but here it is. (See first paragraph).