Updated daily school report generation
A few tweaks were made to covid-warren-school.py. The directory
school-report now contains a script (build-school-report.sh) to combine
a div output of the school report with appropriate markdown
in order to create and updated report. 01-school-report.md is the
preamble text and 03-explanations describes the report itself. The
date the final markdown was generated is added as well.