Test Date Attribution adjustment and wcil daily
The difference between the IDPH data and WCHD data cannot
be fixed by a simple day shift. For this reason I've moved
all date attribution back to the date of the WCHD report.
This means all analysis is of the reports and is up to date
in so much as those reports are timely. Things are usually
pretty close to IDPH. mr238's current report and Historical
report were adjusted accordingly.

I've also pulled together a dashboard (just a bunch of tables)
that will probably continue to develop as a daily county update
that I will host on gh-pages.