WCIL daily dashboard now reports a window
The original dashboard gave daily stats only for the current
week. This creates a bit of a vacuum of information at the
start of the week. I've expanded the daily test reports and
the daily trend measures to a 10 day window to correct for this
and to align with the 10 day window trend metrics.