Moved MR238 reporting and daily update
In addition to the regular daily update, this commit marks the
movement of MR238's metric page to the covid site from my
personal/professional site. This also means there's not a publicly
viewable link (it was shadow live before).

To accommodate this move I needed to update some links and modify
scripts for generating daily sites. This also opens ups the
possibility of a single update script to build new wcil and MR238
daily reports and weekly history updates.