Weekly Report with some Code Updates
In addition to some basic clean up in the weekly reporting script,
I added some code for doing weekly and monthly comparisons to
exploratory.py. This allowed for a top-ten worst weeks comparison
to this week (the current worst) that included a few basic tables
and a week-to-week case breakdown.