Start Transitioning to IDPH daily numbers
The WCHD stopped issuing daily reports in order to
keep up with vaccines. IDPH still issues totals
every day and I can recapture a lot of what I've
reported from that. The big missing piece is
demographics.

wcilviaIDPHDailyDash.py is based off IDPH data.
It's pretty much good to go, but the histogram
and box plot graphics are going to be different
because the data only goes back to 10/18. There's
nothing I can do about that so some comments are
needed. Maybe adjust them for a 3 month range?