WCIL Daily Dash Update
The WCIL daily dashboard now pull only IDPH Data.
To account for the lack of historical data,
the histogram and box-plot graphics now only display
data for the months reported on the daily dashboard.

The weekly and historical dashboards will continue
to use WCHD data as much as possible. IDPH data
will probably be needed for day to day charts.