Vaccine Data Update
I've added daily (with average) vaccine
data to the daily and regional dashboards
and a post to go along with them as well.

The new graph shows daily new full vaccinations
along with a 7 day average. This data is presented
in the same graph as the percent of the population
vaccinated. The regional dashboard scales the daily
counts (and average) on a per 100,000 people basis.